requirements.txt
onnxruntime-gpu==1.18.0
pillow==10.3.0
gdown==5.2.0
numpy==1.26.0
opencv-python-headless==4.10.0.84
ftfy==6.2.0
regex==2024.5.15
scipy==1.13.1
gradio==3.26.0
